requirements-fabric.txt
# AgentiCraft Protocol Fabric Requirements

# Core AgentiCraft
agenticraft>=0.1.0

# Protocol SDKs (when available)
# mcp>=0.1.0                    # Model Context Protocol SDK
# a2a-protocol>=0.1.0           # Google A2A SDK

# Required for protocol implementations
aiohttp>=3.8.0                  # For HTTP-based protocols (ACP)
websockets>=10.0                # For WebSocket connections

# Optional: Integration components
# For memory integration
# redis>=4.0.0                  # If using Redis memory store
# chromadb>=0.3.0               # If using vector memory

# For telemetry integration
opentelemetry-api>=1.15.0       # Telemetry support
opentelemetry-sdk>=1.15.0       # Telemetry implementation
opentelemetry-exporter-prometheus>=0.36b0  # Metrics export

# For security integration
# cryptography>=38.0.0          # For encryption/signing

# For enhanced features
pydantic>=2.0.0                 # Data validation
pyyaml>=6.0                     # YAML configuration
python-dotenv>=1.0.0            # Environment configuration

# Development dependencies
pytest>=7.0.0                   # Testing
pytest-asyncio>=0.21.0          # Async test support
black>=23.0.0                   # Code formatting
mypy>=1.0.0                     # Type checking
